Adult Income Predictions using KNN Machine Learning Algorithm
# UCI Machine Learning Repository - Adult Income Data set

For this project we will be worked with the [Adult Income Data Set](https://archive.ics.uci.edu/ml/datasets/adult). This is a very famous data set and quite many papers have been published on the same!

We'll be trying to predict a classification- income less than equal to 50 or greater than 50
Let's begin our understanding of implementing K-Nearest Neighbour in Python for classification.

We'll use the raw version of the adult income data set, and perform some categorical variables encoding (dummy variables) in python for quite many features as the feautures are mostly String.