Brest Cancer Prediction Project - UCI Data Repository
For this project we will be worked with the UCI Repository Wiconsin Brest Cancer Data Set. This is a very famous data set and quite many papers have been published on the same!

We'll be trying to predict a classification- Diagnosis is Malignant '4' or benign '2'. Let's begin our understanding of implementing K-Nearest Neighbour in Python for classification.

We'll used the raw version of the data set (the one provided), and performed some categorical variables encoding (dummy variables) in python for quite many features as required.