Prostate_Cancer_Project
UCI Machine Learning Repository - Prostate Cancer Data set
For this project we will worked with the UCI Repository Prostate Cancer Data Set. This is a very famous data set and quite many papers have been published on the same!

We'll be tried to predict a classification- Diagnosis is Malignant 'M' or benign 'B'. Let's begin our understanding of implementing K-Nearest Neighbour in Python for classification.

We'll used the raw version of the data set (the one provided), and perform some categorical variables encoding (dummy variables) in python for quite many features as required.