Iris Classification Project
IRIS Data Project
For this project we will be working with the Iris Data Set. This is a very famous data set and quite many papers have been published on the same!

We'll be trying to predict a classification- What type of flower is it. Let's begin our understanding of implementing K-Nearest Neighbour in Python for classification.

We'll use the raw version of the data set (the one provided), and perform some categorical variables encoding (dummy variables) in python for quite many features if required.